deleted mood context and added function in weeklyMood
