+++ src/index.js
@@ -5,22 +5,16 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter as Router } from "react-router-dom";
 import { AuthContextWrapper } from "./context/auth.context";
-import { UserContextWrapper } from "./context/user.context";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
     <AuthContextWrapper>
-      <UserContextWrapper>
-        <Router>
-          <App />
-        </Router>
-      </UserContextWrapper>
+      <Router>
+        <App />
+      </Router>
     </AuthContextWrapper>
   </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
